# doc/results-control-api-gateway/swagger.json
{"swagger":"2.0","info":{"description":"API интерфейс для журнала с оценками и посещаемостью","version":"1.0","title":"Api Documentation","termsOfService":"urn:tos","contact":{},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost","basePath":"/","tags":[{"name":"api-gateway-controller","description":"Api Gateway Controller"},{"name":"basic-error-controller","description":"Basic Error Controller"},{"name":"results-control-api-gateway-application","description":"Results Control Api Gateway Application"}],"paths":{"/auto-attendance/body":{"post":{"tags":["api-gateway-controller"],"summary":"Ð£ÑÑÐ°Ð½Ð¾Ð²ÐºÐ° Ð¿Ð¾ÑÐµÑÐµÐ½Ð¸Ñ ÑÑÑÐ´ÐµÐ½ÑÐ¾Ð¼ Ð·Ð°Ð½ÑÑÐ¸Ñ (ÑÐ½ÑÑÐ¸Ðµ Ð¾ÑÐ¼ÐµÑÐºÐ¸)","description":"ÐÐ»Ñ ÑÑÑÐ°Ð½Ð¾Ð²ÐºÐ¸ Ð¿ÑÐ¸Ð·Ð½Ð°ÐºÐ° Ð¿Ð¾ÑÐµÑÐµÐ½Ð¸Ñ Ð¿Ð°ÑÐ°Ð¼ÐµÑÑ attendanceFact ÑÐ°Ð²Ð½ÑÐ¼ Ð - Ð¸Ð»Ð¸ Ð - Ð´Ð»Ñ Ð¿ÑÐ¾Ð³ÑÐ»Ð° Ð¼Ð¾Ð´ÐµÐ»Ð¸ ResultJson","operationId":"autoAttendanceUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"body","description":"body","required":true,"schema":{"$ref":"#/definitions/AutoAttendanceRequestBodyJson"}}],"responses":{"200":{"description":"OK"}}},"put":{"tags":["api-gateway-controller"],"summary":"Ð£ÑÑÐ°Ð½Ð¾Ð²ÐºÐ° Ð¿Ð¾ÑÐµÑÐµÐ½Ð¸Ñ ÑÑÑÐ´ÐµÐ½ÑÐ¾Ð¼ Ð·Ð°Ð½ÑÑÐ¸Ñ (ÑÐ½ÑÑÐ¸Ðµ Ð¾ÑÐ¼ÐµÑÐºÐ¸)","description":"ÐÐ»Ñ ÑÑÑÐ°Ð½Ð¾Ð²ÐºÐ¸ Ð¿ÑÐ¸Ð·Ð½Ð°ÐºÐ° Ð¿Ð¾ÑÐµÑÐµÐ½Ð¸Ñ Ð¿Ð°ÑÐ°Ð¼ÐµÑÑ attendanceFact ÑÐ°Ð²Ð½ÑÐ¼ Ð - Ð¸Ð»Ð¸ Ð - Ð´Ð»Ñ Ð¿ÑÐ¾Ð³ÑÐ»Ð° Ð¼Ð¾Ð´ÐµÐ»Ð¸ ResultJson","operationId":"autoAttendanceUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"body","description":"body","required":true,"schema":{"$ref":"#/definitions/AutoAttendanceRequestBodyJson"}}],"responses":{"200":{"description":"OK"}}}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"head":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingHEAD","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"post":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPOST","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"put":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPUT","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"delete":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingDELETE","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"options":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingOPTIONS","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}}}},"patch":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPATCH","consumes":["application/json"],"produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}}}}},"/eventResults/{id}":{"get":{"tags":["api-gateway-controller"],"summary":"ÐÑÑÐ½Ð°Ð» Ð¿Ð¾ÑÐµÑÐµÐ½Ð¸Ð¹","description":"ÐÑÐ¾Ð±ÑÐ°Ð¶ÐµÐ½Ð¸Ðµ Ð¶ÑÑÐ½Ð°Ð»Ð° Ð¿Ð¾ÑÐµÑÐµÐ½Ð¸Ñ Ð»ÐµÐºÑÐ¸Ð¸ Ñ Ð¾ÑÐµÐ½ÐºÐ¼Ð¸","operationId":"getEventResultsUsingGET","produces":["*/*"],"parameters":[{"name":"eventId","in":"path","description":"ID ÑÐ¾Ð±ÑÑÐ¸Ñ","required":true,"type":"string","x-example":"initialEvent_1"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ObservableÂ«ListÂ«ResultJsonÂ»Â»"}}}}},"/ping{?value}":{"get":{"tags":["results-control-api-gateway-application"],"summary":"ping","operationId":"pingUsingGET","produces":["*/*"],"parameters":[{"name":"value","in":"query","description":"value","required":false,"type":"string","default":"ok"}],"responses":{"200":{"description":"OK","schema":{"type":"string"}}}}}},"definitions":{"AutoAttendanceRequestBodyJson":{"type":"object","properties":{"attendanceFact":{"type":"string"},"eventId":{"type":"string"},"personId":{"type":"string"}},"title":"AutoAttendanceRequestBodyJson"},"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["ACCEPTED","ALREADY_REPORTED","BAD_GATEWAY","BAD_REQUEST","BANDWIDTH_LIMIT_EXCEEDED","CHECKPOINT","CONFLICT","CONTINUE","CREATED","DESTINATION_LOCKED","EXPECTATION_FAILED","FAILED_DEPENDENCY","FORBIDDEN","FOUND","GATEWAY_TIMEOUT","GONE","HTTP_VERSION_NOT_SUPPORTED","IM_USED","INSUFFICIENT_SPACE_ON_RESOURCE","INSUFFICIENT_STORAGE","INTERNAL_SERVER_ERROR","I_AM_A_TEAPOT","LENGTH_REQUIRED","LOCKED","LOOP_DETECTED","METHOD_FAILURE","METHOD_NOT_ALLOWED","MOVED_PERMANENTLY","MOVED_TEMPORARILY","MULTIPLE_CHOICES","MULTI_STATUS","NETWORK_AUTHENTICATION_REQUIRED","NON_AUTHORITATIVE_INFORMATION","NOT_ACCEPTABLE","NOT_EXTENDED","NOT_FOUND","NOT_IMPLEMENTED","NOT_MODIFIED","NO_CONTENT","OK","PARTIAL_CONTENT","PAYLOAD_TOO_LARGE","PAYMENT_REQUIRED","PERMANENT_REDIRECT","PRECONDITION_FAILED","PRECONDITION_REQUIRED","PROCESSING","PROXY_AUTHENTICATION_REQUIRED","REQUESTED_RANGE_NOT_SATISFIABLE","REQUEST_ENTITY_TOO_LARGE","REQUEST_HEADER_FIELDS_TOO_LARGE","REQUEST_TIMEOUT","REQUEST_URI_TOO_LONG","RESET_CONTENT","SEE_OTHER","SERVICE_UNAVAILABLE","SWITCHING_PROTOCOLS","TEMPORARY_REDIRECT","TOO_EARLY","TOO_MANY_REQUESTS","UNAUTHORIZED","UNAVAILABLE_FOR_LEGAL_REASONS","UNPROCESSABLE_ENTITY","UNSUPPORTED_MEDIA_TYPE","UPGRADE_REQUIRED","URI_TOO_LONG","USE_PROXY","VARIANT_ALSO_NEGOTIATES"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}},"title":"ModelAndView"},"ObservableÂ«ListÂ«ResultJsonÂ»Â»":{"type":"object","title":"ObservableÂ«ListÂ«ResultJsonÂ»Â»"},"ResultJson":{"type":"object","properties":{"createdAt":{"type":"string","format":"date-time"},"lessionId":{"type":"object","example":"initialLesson_1","description":"ID ÑÑÐ¾ÐºÐ°"},"lessionName":{"type":"string"},"personId":{"type":"string"},"personName":{"type":"string"},"resultType":{"type":"string"},"value":{"type":"string"}},"title":"ResultJson"},"View":{"type":"object","properties":{"contentType":{"type":"string"}},"title":"View"}}}